requirements.txt
fastapi
pyyaml
markdown
uvicorn
jinja2
watchdog